+++ index.html
@@ -25,7 +25,7 @@
     <button id="game-restart-button" hidden>재시작</button>
   </div>
   <script src="https://cdn.jsdelivr.net/npm/@woowacourse/mission-utils@1.0.1/dist/mission-utils.min.js"></script>
-  <script type="module" src="/src/index.js"></script>
+  <script type="module" src="src/index.js"></script>
 </body>
 
 </html>
